refactor(DamFooter): split up UploadFooter and SelectionFooter
